msibuilder.ps1: Fix WiX download
Use our own copy of wix314.exe since the official download
servers have been down for days.

Remove the weird Start-Job work-around while here. For reasons
I do not understand, it suddenly broke for me. But the failure
occured only when running it via packer and only on first run,
not when retrying the step. Weird stuff. Anyway, replace the
work-around by checking the return code of the installation to
make sure it can't fail silently.

Signed-off-by: Frank Lichtenheld <frank@lichtenheld.com>