Allow canceling old builds when new commits have arrived
This is meant for development buildbot systems that are not always up. On those
systems, when buildmaster is launched, it often starts running builds for some
obsolete and completely unrelevant commit, which may take a long, long time to
finish. With cancel_old_builds set to true in master.ini a new commit to a
non-master branch will cancel all previous builds.

Signed-off-by: Samuli Seppänen <samuli.seppanen@gmail.com>