buildmaster/openvpn: Make builds parallel
Since the actual build is so small and instead
much of the time is spent in steps like
autoconf/configure/tests that are not CPU-bound.
So parallelizing the code steps is possible
even on shared resources like the docker workers.

For the t_client step the -j2 makes a big difference
since the two minute t_cltsrv.sh can be run in
parallel to t_client.sh.

Signed-off-by: Frank Lichtenheld <frank@lichtenheld.com>